사운드 오프셋 옵션 추가

+++ Assets/Scripts/Globals/GameSetting.cs
@@ -77,7 +77,7 @@ public class GameSetting : SingletonUnity<GameSetting>
     public static double PreLoadTime => ( 1500d / Weight );
 
     // Sound
-    public static double SoundOffset = 0d;
+    public static int SoundOffset = 0;
 
     // Opacity Percentage ( 0 ~ 100 )
     public static float BGAOpacity   = 100f;
